Allow Tenant Impersonation
This change allows for impersonating in to target tenants via Addigy's partner portal internal API.

To use it, call `getAuthToken()` to retrieve an authentication token into your top-level tenant. From there, pass that auth token and an organization ID to `getImpersonationToken()`, which will then return back a new authentication token for use on subsequent calls to the child tenant.